C, parse attributes
